Constants and Types

Constants and Types

The following constants are defined in ModiaMath (all these definitions are not exported and therefore need to be prefixed with ModiaMath):

const path
Absolute path of ModiaMath package directory. This allows for example to run ModiaMath examples as include("$(ModiaMath.path)/examples/Simulate_Pendulum.jl").

const Version
Version of ModiaMath

const Date
Version Date of ModiaMath

The following abstract types are defined and used in ModiaMath:

ModiaMath.AbstractComponentInternal — Type.

The internal part of a component (has at least fields "name" and "within")

source

ModiaMath.AbstractComponentWithVariables — Type.

type ModiaMath.AbstractComponentWithVariables

Struct that contains ModiaMath.AbstractVariables as field or as field in a sub-struct.

source

ModiaMath.AbstractRealVariable — Type.

ModiaMath.AbstractRealVariable <: ModiaMath.AbstractVariable

A real ModiaMath.AbstractVariable (either scalar or array)

source

ModiaMath.AbstractSimulationModel — Type.

type ModiaMath.AbstractSimulationModel

Struct that is used as simulation model (has field: simulationState)

source

ModiaMath.AbstractVariable — Type.

type ModiaMath.AbstractVariable <: ModiaMath.AbstractComponentWithVariables

A Variable used as element of the DAE model description and is included in the result (if no residue)

source