DAE

DAE

ModiaMath.DAE — Module.

module ModiaMath.DAE

Interface between the ModiaMath.SimulationEngine and the index 1 DAE model.

The following functions can be called in the DAE model to inquire information about the simulation state:

ModiaMath.getTime

ModiaMath.getStartTime

ModiaMath.getStopTime

ModiaMath.getTolerance

ModiaMath.isInitial

ModiaMath.isTerminal

ModiaMath.isEvent

ModiaMath.isZeroCrossing

ModiaMath.isAfterSimulationStart

ModiaMath.isStoreResult

ModiaMath.isLogInfos

ModiaMath.isLogWarnings

ModiaMath.isLogEvents

The following functions can be called in the DAE model to set properties in the simulation engine:

ModiaMath.setNominal!

ModiaMath.setNextEvent!

ModiaMath.positive!

ModiaMath.negative!

ModiaMath.change!

ModiaMath.edge!

The following functions can be either called in the DAE model or they can be called on a simulation model (before or after ModiaMath.simulate!(simulationModel, ...) is called).

ModiaMath.logOn!

ModiaMath.logOff!

ModiaMath.setLogCategories!

Main developer

Martin Otter, DLR - Institute of System Dynamics and Control

source

ModiaMath.DAE.EventRestart — Type.

@enum EventRestart NoRestart Restart FullRestart Terminate

Define how to continue or restart integration after an event. Usually, Restart should be used. Only in special cases, the other flags are useful.

NoRestart, continue integration without restarting the integrator

Restart, restart integrator

FullRestart, restart integrator and optionally exchange simulationState (so dimensions may change)

Terminate, terminate integration

source

ModiaMath.DAE.change! — Method.

ModiaMath.change!(sim, nr, crossing, crossingAsString; restart = ModiaMath.Restart)

Trigger an event, whenever crossing > 0 changes from false to true or from true to false. The function returns always false.

source

ModiaMath.DAE.edge! — Method.

ModiaMath.edge!(sim, nr, crossing, crossingAsString; restart = ModiaMath.Restart)

Trigger an event, whenever crossing > 0 switches from false to true. The function returns always false.

source

ModiaMath.DAE.getStartTime — Method.

ModiaMath.getStartTime(m::ModiaMath.[AbstractSimulationModel|SimulationState])

Return startTime of the actual simulation run.

source

ModiaMath.DAE.getStopTime — Method.

ModiaMath.getStopTime(m::ModiaMath.[AbstractSimulationModel|SimulationState])

Return stopTime of the actual simulation run (when the simulation will be terminated).

source

ModiaMath.DAE.getTime — Method.

ModiaMath.getTime(m::ModiaMath.[AbstractSimulationModel|SimulationState])

Return actual simulation time.

source

ModiaMath.DAE.getTolerance — Method.

ModiaMath.getTolerance(m::ModiaMath.[AbstractSimulationModel|SimulationState])

Return (relative) tolerance of the actual simulation run.

source

ModiaMath.DAE.isAfterSimulationStart — Method.

ModiaMath.isAfterSimulationStart(m::ModiaMath.[AbstractSimulationModel|SimulationState])

Return true, if after start of simulation and false if during initialization.

source

ModiaMath.DAE.isEvent — Method.

ModiaMath.isEvent(m::ModiaMath.[AbstractSimulationModel|SimulationState])

Return true, if event phase of simulation (including initialization)

source

ModiaMath.DAE.isInitial — Method.

ModiaMath.isInitial(m::ModiaMath.[AbstractSimulationModel|SimulationState])

Return true, if initialization phase of simulation.

source

ModiaMath.DAE.isStoreResult — Method.

ModiaMath.isStoreResult(m::ModiaMath.[AbstractSimulationModel|SimulationState])

Return true, if communication point and variables shall be stored in the result data structure.

source

ModiaMath.DAE.isTerminal — Method.

ModiaMath.isTerminal(m::ModiaMath.[AbstractSimulationModel|SimulationState])

Return true, if terminal phase of simulation.

source

ModiaMath.DAE.isZeroCrossing — Method.

ModiaMath.isZeroCrossing(m::ModiaMath.[AbstractSimulationModel|SimulationState])

Return true, if simulation model shall compute zero-crossing functions (as required by the integrator).

source

ModiaMath.DAE.negative! — Method.

ModiaMath.negative!(sim, nr, crossing, crossingAsString; restart = ModiaMath.Restart)

Return crossing < 0 such that a state event is triggered whenever crossing < 0 changes its value.

source

ModiaMath.DAE.positive! — Method.

ModiaMath.positive!(sim, nr, crossing, crossingAsString; restart = ModiaMath.Restart)

Return crossing > 0 such that a state event is triggered whenever crossing > 0 changes its value.

Note, a small hysteresis is added to the crossing function during continuous integration, in order to reduce issues during event iterations due to small numerical errors. However, at an event instant, crossing > 0 is returned without any hysteresis.

Arguments

sim::ModiaMath.SimulationState: (Internal) simulation state provided by simulation engine.

nr::Int: (> 0 required) Every call of positive!(..) must be identified by a unique value of nr This value is used as index in a vector that holds the internal memory for crossing > 0.

crossing::Float64: Zero crossing function.

crossingAsString::String: crossing as string representation. This string is used for log messages.

restart::ModiaMath.EventRestart: Restart action after the crossing > 0 event occurred.

Example

function computeVariables!(m::Model, sim::SimulationState)
   ...
   # f = s > 0 ? fMax : 0.0
   m.sPos.value = ModiaMath.positive!(sim, 1, m.s.value, "s")
   m.f.value    = m.sPos.value ? m.fMax.value : 0.0
   ...
end

source

ModiaMath.DAE.setNextEvent! — Method.

ModiaMath.setNextEvent!(sim, nextEventTime; integratoToEvent=true, restart = ModiaMath.Restart)

At an event instant (isEvent(sim) == true) trigger the next time event.

Arguments

sim::ModiaMath.SimulationState: (Internal) simulation state provided by simulation engine.

nextEventTime::Float64: Time instant of the next time event.

integrateToEvent::Bool: If true, the integrator integrates exactly to this event. If false, the integrator might integrate beyond nextEventTime (so the step size of the integrator is not influenced by this event). This option is only useful, if information is inquired at the event and restart=ModiaMath.NoRestart is used. The benefit is that the integrator is practically not influenced by this event.

restart::ModiaMath.EventRestart: Restart action after the event at which setNextEvent! was called.

source

ModiaMath.DAE.setNominal! — Method.

ModiaMath.setNominal!(sim::ModiaMath.SimulationState, x_nominal::Vector{Float64})

At initialization provide x_nominal, the nominal values of the x-vector, and store them insim`, the simulation state. These values are used to compute the absolute tolerances of the x-vector for the integrator.

Example

function getModelResidues!(m::Model, t, x, derx, r, w)
   sim = m.simulationState
   if ModiaMath.isInitial(sim)
      ModiaMath.setNominal!(sim, [1.0, 1e-5, 1e8])   # if length(x)=3
   end
   ...
end

source